intern fhir::Sort
